Rename decorators to transformers
